# packages/contracts/openapi/nova-file-api.openapi.json
{
  "components": {
    "responses": {
...
      "FileInvalidRequestResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        },
        "description": "Canonical invalid-request response."
      },
...
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "additionalProperties": true,
            "properties": {
              "code": {
                "type": "string"
              },
              "details": {
                "type": "object"
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ResourcePlanItem": {
        "additionalProperties": false,
        "description": "Resource planning decision per requested resource.",
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "resource": {
            "title": "Resource",
            "type": "string"
          },
          "supported": {
            "title": "Supported",
            "type": "boolean"
          }
        },
        "required": [
          "resource",
          "supported"
        ],
        "title": "ResourcePlanItem",
        "type": "object"
      },
      "ResourcePlanRequest": {
        "additionalProperties": false,
        "description": "Resource planning request body.",
        "properties": {
          "resources": {
            "items": {
              "maxLength": 256,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 256,
            "minItems": 1,
            "title": "Resources",
            "type": "array"
          }
        },
        "required": [
          "resources"
        ],
        "title": "ResourcePlanRequest",
        "type": "object"
      },
      "ResourcePlanResponse": {
        "additionalProperties": false,
        "description": "Resource planning response body.",
        "properties": {
          "plan": {
            "items": {
              "$ref": "#/components/schemas/ResourcePlanItem"
            },
            "maxItems": 256,
            "title": "Plan",
            "type": "array"
          }
        },
        "required": [
          "plan"
        ],
        "title": "ResourcePlanResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "X-Worker-Token": {
        "description": "Worker token header for trusted job-worker calls.",
        "in": "header",
        "name": "X-Worker-Token",
        "type": "apiKey"
      },
      "sessionAuth": {
        "description": "Session identifier header for caller context.",
        "in": "header",
        "name": "X-Session-Id",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "nova-file-api",
    "version": "0.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/resources/plan": {
      "post": {
        "description": "Plan supportability for requested resource keys.",
        "operationId": "plan_resources",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourcePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourcePlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/FileInvalidRequestResponse"
          }
        },
        "summary": "Plan Resources",
        "tags": [
          "platform"
        ]
      }
    },
...
  }
}